Common IRepository interface.

+++ src/EFPersistenceModel/ArticleWordRepository.cs
@@ -1,13 +1,14 @@
 ﻿namespace ArticleHarbor.EFPersistenceModel
 {
     using System;
+    using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
     using ArticleHarbor.DomainModel;
     using ArticleHarbor.EFDataAccess;
     using ArticleWord = DomainModel.ArticleWord;
 
-    public class ArticleWordRepository : IArticleWordRepository
+    public class ArticleWordRepository : IRepository<ArticleWord>
     {
         private readonly ArticleHarborDbContext context;
 
@@ -33,21 +34,45 @@
             return articleWord == null ? null : articleWord.ToDomain();
         }
 
-        public Task InsertAsync(ArticleWord articleWord)
+        public Task<ArticleWord> InsertAsync(ArticleWord item)
         {
-            if (articleWord == null)
-                throw new ArgumentNullException("articleWord");
+            if (item == null)
+                throw new ArgumentNullException("item");
 
-            if (this.Select(articleWord.ArticleId, articleWord.Word) == null)
-                this.context.ArticleWords.Add(articleWord.ToPersistence());
+            if (this.Select(item.ArticleId, item.Word) == null)
+                this.context.ArticleWords.Add(item.ToPersistence());
 
-            return Task.FromResult<object>(null);
+            return Task.FromResult<ArticleWord>(item);
         }
 
-        public Task DeleteAsync(int articleId)
+        public Task<ArticleWord> FineAsync(params object[] identity)
         {
+            if (identity == null)
+                throw new ArgumentNullException("identity");
+
+            throw new NotImplementedException();
+        }
+
+        public Task<IEnumerable<ArticleWord>> SelectAsync()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task UpdateAsync(ArticleWord item)
+        {
+            if (item == null)
+                throw new ArgumentNullException("item");
+            throw new NotImplementedException();
+        }
+
+        public Task DeleteAsync(params object[] identity)
+        {
+            if (identity == null)
+                throw new ArgumentNullException("identity");
+
+            var id = (int)identity[0];
             var articleWords = this.context.ArticleWords
-                .Where(x => x.ArticleId == articleId).ToArray();
+                .Where(x => x.ArticleId == id).ToArray();
 
             foreach (var articleWord in articleWords)
                 this.context.ArticleWords.Remove(articleWord);
